move VisitLogRestController into EventRestContriller
