+++ restservice/src/test/java/org/itevents/controller/EventRestControllerTest.java
@@ -3,8 +3,11 @@ package org.itevents.controller;
 import org.codehaus.jackson.map.ObjectMapper;
 import org.itevents.model.Event;
 import org.itevents.model.User;
+import org.itevents.model.VisitLog;
+import org.itevents.model.builder.VisitLogBuilder;
 import org.itevents.service.EventService;
 import org.itevents.service.UserService;
+import org.itevents.service.VisitLogService;
 import org.itevents.test_utils.BuilderUtil;
 import org.junit.Before;
 import org.junit.Test;
@@ -14,8 +17,8 @@ import org.mockito.Mock;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -26,6 +29,8 @@ public class EventRestControllerTest extends AbstractControllerTest {
     private EventService eventService;
     @Mock
     private UserService userService;
+    @Mock
+    private VisitLogService visitLogService;
     @InjectMocks
     private EventRestController eventRestController;
 
@@ -86,4 +91,25 @@ public class EventRestControllerTest extends AbstractControllerTest {
                 .andExpect(status().isOk())
                 .andExpect(content().json(expectedUsersInJson));
     }
+
+    @Test
+    public void shouldAddVisitLogToEventJavaWithAnonymous() throws Exception {
+        Event event = BuilderUtil.buildEventJava();
+        User userGuest = BuilderUtil.buildUserGuest();
+        VisitLog visitLog = VisitLogBuilder.aVisitLog().event(event).user(userGuest).build();
+
+        when(eventService.getEvent(event.getId())).thenReturn(event);
+        when(userService.getAuthorizedUser()).thenReturn(userGuest);
+        doNothing().when(visitLogService).addVisitLog(visitLog);
+
+        mockMvc.perform(get("/events/" + event.getId() + "/register"))
+                .andExpect(content().string(event.getRegLink()))
+                .andExpect(status().isOk());
+
+        verify(eventService).getEvent(event.getId());
+        verify(userService).getAuthorizedUser();
+        verify(visitLogService).addVisitLog(any(VisitLog.class));
+    }
+
+
 }
